fix: icon color
