+++ packages/nc-gui/components/dashboard/HomeSidebar/index.vue
@@ -1,11 +1,25 @@
 <script setup lang="ts">
 const { user, isMobileMode } = useGlobal()
 
+const router = useRouter()
+
+const route = router.currentRoute
+
+const { t } = useI18n()
+
 const workspaceStore = useWorkspace()
 
-const { activeWorkspaceId } = storeToRefs(workspaceStore)
+const { activeWorkspaceId, activeWorkspace, workspaceUserCount } = storeToRefs(workspaceStore)
+const { loadCollaborators } = workspaceStore
+
+const basesStore = useBases()
+const { isProjectsLoaded, basesList } = storeToRefs(basesStore)
 
 const { isLeftSidebarOpen } = storeToRefs(useSidebarStore())
+
+const { isUIAllowed } = useRoles()
+
+const { wsTabVisibility } = useWorkspaceTabVisibility(activeWorkspace)
 
 const notificationStore = useNotification()
 
@@ -13,16 +27,60 @@ const { unreadCount } = toRefs(notificationStore)
 
 const { isDark } = useTheme()
 
-const name = computed(() => user.value?.display_name?.trim())
-
 const isUserMenuOpen = ref(false)
 
 const isNotificationOpen = ref(false)
 
-const searchQuery = useState<string>('ws-home-search', () => '')
+const name = computed(() => user.value?.display_name?.trim())
 
-const navigateToWorkspace = () => {
-  navigateTo(`/${activeWorkspaceId.value}`)
+// ── Sidebar navigation ──
+
+interface NavItem {
+  key: string
+  icon: string
+  label: string
+  count?: number
+  hidden?: boolean
+}
+
+const navItems = computed<NavItem[]>(() => {
+  return [
+    {
+      key: 'bases',
+      icon: 'ncBaseOutline',
+      label: t('objects.projects'),
+      count: isProjectsLoaded.value ? basesList.value.length : undefined,
+    },
+    {
+      key: 'collaborators',
+      icon: 'users',
+      label: t('labels.inviteMembers'),
+      count: workspaceUserCount.value,
+      hidden: !wsTabVisibility.value.collaborators,
+    },
+    {
+      key: 'integrations',
+      icon: 'integration',
+      label: t('general.integrations'),
+      hidden: !wsTabVisibility.value.integrations,
+    },
+  ].filter((item) => !item.hidden)
+})
+
+const activeNavKey = computed(() => {
+  if (isWsAdminRoute(route.value)) return 'admin'
+
+  return routeNameToWsTab[route.value.name as string] || 'bases'
+})
+
+function onNavClick(item: NavItem) {
+  if (item.key === 'collaborators' && isUIAllowed('workspaceCollaborators')) {
+    loadCollaborators({}, activeWorkspaceId.value)
+  }
+
+  const typeOrId = route.value.params.typeOrId || activeWorkspaceId.value || 'nc'
+
+  router.push({ name: wsTabToRouteName[item.key] || 'index-typeOrId', params: { typeOrId } })
 
   if (isMobileMode.value) {
     isLeftSidebarOpen.value = false
@@ -31,49 +89,39 @@ const navigateToWorkspace = () => {
 </script>
 
 <template>
-  <div class="nc-home-sidebar flex flex-col h-full bg-nc-bg-gray-sidebar border-r-1 border-nc-border-gray-light select-none">
+  <div class="nc-home-sidebar flex flex-col h-full bg-nc-bg-default select-none" style="--topbar-height: 3.5rem">
     <!-- Brand header -->
-    <div class="w-full px-2 py-1.5 flex items-center justify-between gap-2 h-[var(--topbar-height)] flex-none">
+    <div
+      class="w-full px-2 py-1.5 flex items-center justify-between gap-2 h-[var(--topbar-height)] flex-none border-b-1 border-nc-border-gray-light"
+    >
       <div class="pl-1">
         <img v-if="isDark" alt="NocoDB" src="~/assets/img/brand/full-logo.png" class="h-9" />
         <img v-else alt="NocoDB" src="~/assets/img/brand/nocodb-full-color.png" class="h-9" />
       </div>
-
-      <GeneralHideLeftSidebarBtn show-always />
     </div>
 
-    <!-- Search input (desktop only — mobile has search in base list header) -->
-    <div class="hidden md:flex px-2 h-[var(--toolbar-height)] items-center">
-      <a-input
-        v-model:value="searchQuery"
-        :placeholder="$t('activity.searchProject')"
-        allow-clear
-        class="nc-input-border-on-value nc-home-sidebar-search nc-input-sm"
-      >
-        <template #prefix>
-          <GeneralIcon icon="search" class="text-nc-content-gray-muted/80 mr-0.5" />
-        </template>
-      </a-input>
-    </div>
-
-    <!-- Workspace section -->
-    <div class="flex-1 flex flex-col overflow-hidden nc-project-home-section !pb-0">
+    <!-- Navigation section -->
+    <div class="flex-1 flex flex-col overflow-hidden pt-2">
       <div class="nc-ws-section-header flex items-center justify-between">
         <span>{{ $t('objects.workspace') }}</span>
       </div>
 
-      <div class="flex-1 overflow-y-auto nc-scrollbar-thin px-1">
+      <div class="flex-1 overflow-y-auto nc-scrollbar-thin px-2">
         <NcSidebarMenuItem
-          v-if="activeWorkspaceId"
-          class="group !my-1 !h-11 !gap-3 !text-sm"
-          :active="true"
-          data-testid="nc-home-sidebar-ws-nc"
-          @click="navigateToWorkspace()"
+          v-for="item in navItems"
+          :key="item.key"
+          :icon="item.icon"
+          :active="activeNavKey === item.key"
+          class="!h-8 !my-0.5"
+          :data-testid="`nc-ws-sidebar-${item.key}`"
+          @click="onNavClick(item)"
         >
-          <template #icon>
-            <GeneralIcon icon="ncWorkspace" class="flex-none h-5 w-5" />
+          {{ item.label }}
+          <template #extraRight>
+            <span v-if="item.count !== undefined" class="text-bodySm text-nc-content-gray-muted mr-1.5">
+              {{ item.count }}
+            </span>
           </template>
-          <span class="capitalize">Default Workspace</span>
         </NcSidebarMenuItem>
       </div>
     </div>
@@ -83,7 +131,7 @@ const navigateToWorkspace = () => {
       <div class="flex items-center gap-0.5">
         <NcDropdown v-model:visible="isUserMenuOpen" placement="topLeft" overlay-class-name="!min-w-56">
           <div
-            class="flex items-center gap-2 pl-1.5 pr-2 h-8 rounded-md cursor-pointer flex-1 min-w-0 transition-colors"
+            class="flex items-center gap-2 pl-1.5 pr-2 py-1 rounded-md cursor-pointer flex-1 min-w-0 transition-colors"
             :class="{
               'bg-nc-bg-gray-medium': isUserMenuOpen,
               'hover:bg-nc-bg-gray-medium': !isUserMenuOpen,
@@ -96,6 +144,10 @@ const navigateToWorkspace = () => {
               <NcTooltip show-on-truncate-only class="truncate text-bodyDefaultSm text-nc-content-gray block">
                 <template #title>{{ name || user?.email }}</template>
                 {{ name || user?.email }}
+              </NcTooltip>
+              <NcTooltip v-if="name" show-on-truncate-only class="truncate text-captionSm text-nc-content-gray-muted block">
+                <template #title>{{ user?.email }}</template>
+                {{ user?.email }}
               </NcTooltip>
             </div>
           </div>
@@ -131,6 +183,13 @@ const navigateToWorkspace = () => {
   </div>
 </template>
 
+<style lang="scss">
+// Match the pane splitter to the topbar's bottom border on the ws-home
+.nc-sidebar-content-resizable-wrapper:has(.nc-home-sidebar) > .splitpanes__splitter:before {
+  @apply !bg-nc-border-gray-light;
+}
+</style>
+
 <style lang="scss" scoped>
 .nc-home-sidebar {
   @apply !pb-0;
@@ -138,19 +197,8 @@ const navigateToWorkspace = () => {
 }
 
 .nc-ws-section-header {
-  @apply px-2 pt-1.5 pb-1 font-semibold text-nc-content-brand uppercase tracking-wide;
-  font-size: 13px;
-}
-
-.nc-home-sidebar-search {
-  @apply !rounded-lg;
-
-  :deep(.ant-input) {
-    @apply !border-none !shadow-none !text-bodyDefaultSm;
-  }
-
-  :deep(.ant-input-affix-wrapper) {
-    @apply !border-none !shadow-none rounded-lg px-2 py-1;
-  }
+  @apply pl-5 pr-2 pt-1.5 pb-1.5 font-semibold text-nc-content-gray-muted uppercase;
+  font-size: 11px;
+  letter-spacing: 0.05em;
 }
 </style>
