+++ packages/nc-gui/components/workspace/View.vue
@@ -30,10 +30,6 @@ const { orgId, org } = storeToRefs(orgStore)
 
 const { isWsAuditEnabled, handleUpgradePlan, blockTeamsManagement, showUpgradeToUseTeams } = useEeConfig()
 
-const { isFeatureEnabled } = useBetaFeatureToggle()
-
-const isWsLandingRedesign = computed(() => isFeatureEnabled(FEATURE_FLAG.WORKSPACE_LANDING_REDESIGN))
-
 const { isFromIntegrationPage, eventBus, searchQuery: storeSearchQuery, loadIntegrations } = useProvideIntegrationViewStore()
 
 // Local ref for integrations view mode (main page vs all-connections page).
@@ -300,11 +296,7 @@ if (!props.isNewWsPage) {
           </template>
           <div
             class="nc-integrations-layout nc-content-max-w mx-auto"
-            :class="
-              isNewWsPage && isWsLandingRedesign
-                ? 'h-[calc(100vh-var(--topbar-height))]'
-                : 'h-[calc(100vh-var(--topbar-height)-44px)]'
-            "
+            :class="isNewWsPage ? 'h-[calc(100vh-var(--topbar-height))]' : 'h-[calc(100vh-var(--topbar-height)-44px)]'"
           >
             <!-- Main integrations page -->
             <template v-if="integrationsViewMode === 'main'">
